+++ tests/const.py
@@ -19,7 +19,7 @@ USERNAME = "user"
 SYSVAR_DATA: list[SystemVariableData] = [
     SystemVariableData(
         vid="1",
-        name="alarm",
+        legacy_name="alarm",
         description="",
         data_type=SysvarType.ALARM,
         unit=None,
@@ -31,7 +31,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="2",
-        name="alarm_ext",
+        legacy_name="alarm_ext",
         description="",
         data_type=SysvarType.ALARM,
         unit=None,
@@ -43,7 +43,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="3",
-        name="logic",
+        legacy_name="logic",
         description="",
         data_type=SysvarType.LOGIC,
         unit=None,
@@ -55,7 +55,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="4",
-        name="logic_ext",
+        legacy_name="logic_ext",
         description="",
         data_type=SysvarType.LOGIC,
         unit=None,
@@ -67,7 +67,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="5",
-        name="list",
+        legacy_name="list",
         description="",
         data_type=SysvarType.LIST,
         unit=None,
@@ -79,7 +79,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="6",
-        name="list_ext",
+        legacy_name="list_ext",
         description="",
         data_type=SysvarType.LIST,
         unit=None,
@@ -91,7 +91,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="7",
-        name="string",
+        legacy_name="string",
         description="",
         data_type=SysvarType.STRING,
         unit=None,
@@ -103,7 +103,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="8",
-        name="string_ext",
+        legacy_name="string_ext",
         description="",
         data_type=SysvarType.STRING,
         unit=None,
@@ -115,7 +115,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="9",
-        name="float",
+        legacy_name="float",
         description="",
         data_type=SysvarType.FLOAT,
         unit=None,
@@ -127,7 +127,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="10",
-        name="float_ext",
+        legacy_name="float_ext",
         description="",
         data_type=SysvarType.FLOAT,
         unit="°C",
@@ -139,7 +139,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="11",
-        name="integer",
+        legacy_name="integer",
         description="",
         data_type=SysvarType.INTEGER,
         unit=None,
@@ -151,7 +151,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
     ),
     SystemVariableData(
         vid="12",
-        name="integer_ext",
+        legacy_name="integer_ext",
         description="",
         data_type=SysvarType.INTEGER,
         unit=None,
@@ -165,7 +165,7 @@ SYSVAR_DATA: list[SystemVariableData] = [
 
 PROGRAM_DATA: list[ProgramData] = [
     ProgramData(
-        name="p1",
+        legacy_name="p1",
         pid="pid1",
         description="",
         is_active=True,
@@ -173,7 +173,7 @@ PROGRAM_DATA: list[ProgramData] = [
         last_execute_time="",
     ),
     ProgramData(
-        name="p_2",
+        legacy_name="p_2",
         pid="pid2",
         description="",
         is_active=False,